html/editing/dnd/drop/events-contenteditable-manual.tentative.html

<!doctype html>
<meta charset=utf-8>
<title>Selection drag and drop: events for contenteditable</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
[data-placeholder]:empty::before { content: attr(data-placeholder); }
</style>
<div><input id=a data-select="0,7" value="Drag me" data-expected-events="beforeinput,input"></div>
<div contenteditable=true id=b data-placeholder="...to here"  data-expected-events="drop,beforeinput,textInput,input"></div>
<script src="/uievents/textInput/support/common.js"></script>
<script src="support/events.js"></script>
